Allow to start an auth session from a UI plugin
Currently we have two kinds of providers: Those that do oauth based on static entries in the provider file (e.g. Google) and those with UI plugins that ask for a username+password.

However there are cases where we need to combine the two approaches.

For example for Mastodon we need to first prompt the user to enter an instance URL, make an API call to that instance to register ourselves as application and then start the oauth process based on information we got from the previous API call.

This adds a way for a plugin to trigger an auth session at any time and pass arbitrary data to it